Fixed CMake
