+++ include/LimitOrderBook.h
@@ -24,47 +24,48 @@ struct Order {
 
 // Define the Limit class
 class Limit {
-   public:
-    Limit();
-    Limit(double price);
-    double getPrice();
-    int getTotalVolume();
-    void addOrder(Order order);
-    void removeOrder(int oid);
-    Order getOrder(int oid);
-    Order getPriorityOrder();
-    void updateOrder(int oid, int volume);
+    public:
+        Limit();
+        Limit(double price);
+        double getPrice();
+        int getTotalVolume();
+        void addOrder(Order order);
+        void removeOrder(int oid);
+        Order getOrder(int oid);
+        Order getPriorityOrder();
+        void updateOrder(int oid, int volume);
 
-   private:
-    std::deque<Order> orders;
-    double price;
-    int totalVolume;
+    private:
+        std::deque<Order> orders;
+        double price;
+        int totalVolume;
 };
 
 // Define the LimitOrderBook class
 class LimitOrderBook {
-   public:
-    LimitOrderBook(double max_price, double tick);
-    void addOrder(Dir direction, double price, int volume, int cid);
-    void cancelOrder(int oid);
-    Order getOrder(int oid);
-    double getBestBid();
-    double getBestAsk();
-    void printOrderBook();
+    public:
+        LimitOrderBook(double max_price, double tick);
+        void addOrder(Dir direction, double price, int volume, int cid);
+        void cancelOrder(int oid);
+        Order getOrder(int oid);
+        double getBestBid();
+        double getBestAsk();
+        void printOrderBook();
 
-   private:
-    std::vector<Limit> bids;
-    std::vector<Limit> asks;
-    std::unordered_map<int, Limit*> orderIDs;
-    int nextOID;
-    double tick;
-    double bestBid;
-    double bestAsk;
-    double maxPrice;
+    private:
+        std::vector<Limit> bids;
+        std::vector<Limit> asks;
+        std::unordered_map<int, Limit*> orderIDs;
+        
+        int nextOID;
+        double tick;
+        double bestBid;
+        double bestAsk;
+        double maxPrice;
 
-    void matchBidOrder(Order* order);
-    void matchAskOrder(Order* order);
-    void deleteOrder(int oid);
+        void matchBidOrder(Order* order);
+        void matchAskOrder(Order* order);
+        void deleteOrder(int oid);
 };
 
 #endif  // LIMITORDERBOOK_H
